+++ include/qpp/classes/qbase_engine.hpp
@@ -124,13 +124,16 @@ class QBaseEngine : public IQEngineTraits, public IDisplay, public IJSON {
      * \note This function is a no-op in qpp::QBaseEngine; override it in all
      * derived classes to achieve the desired behaviour
      *
-     * \param elem Step to be executed
+     * \param it Iterator to the step to be executed
      * \return Reference to the current instance
      */
-    virtual QBaseEngine& execute(
-        [[maybe_unused]] const typename QCircuitTraits<QCT>::value_type& elem) {
+    virtual QBaseEngine&
+    execute([[maybe_unused]] typename QCircuitTraits<QCT>::iterator_type it) {
         return *this;
     };
+
+    QBaseEngine& execute([[maybe_unused]]
+                         typename QCircuitTraits<QCT>::value_type step) {}
 
     /**
      * \brief Executes the entire quantum circuit description
@@ -163,19 +166,6 @@ class QBaseEngine : public IQEngineTraits, public IDisplay, public IJSON {
      */
     virtual QBaseEngine& set_state([[maybe_unused]] const T& state) {
         return *this;
-    }
-
-    /**
-     * \brief Executes one step in the quantum circuit description
-     *
-     * \note Do not override!
-     *
-     * \param it Iterator to the step to be executed
-     * \return Reference to the current instance
-     */
-    QBaseEngine&
-    execute(const typename QCircuitTraits<QCT>::iterator_type& it) {
-        return this->execute(*it);
     }
 
     /**