+++ include/qpp/internal/classes/qcircuit_conditional_step.hpp
@@ -92,42 +92,95 @@ struct QCircuitConditionalStep : IDisplay {
     using cond_func_t = std::function<bool(std::vector<idx>)>;
 
     /**
-     * \class qpp::internal::QCircuitConditionalStep::IfElseLoc
-     * \brief Stores the location of conditional statements
+     * \class qpp::internal::QCircuitConditionalStep::Context
+     * \brief Keeps track of the location of conditional statements
      */
-    struct Context {
+    struct Context : IDisplay {
+        idx dit_shift = 0; ///< relative position of the first dit w.r.t. 0,
+                           ///< used when composing circuits
         std::optional<std::pair<idx, cond_func_t>>
             if_expr; ///< location of if statement and corresponding condition
                      ///< function
         std::optional<idx> else_expr;  ///< location of else statement
         std::optional<idx> endif_expr; ///< location of endif statement
+
+        /**
+         * \brief Equality operator
+         * \note Does not test equality of the comparison functions
+         *
+         * \param rhs Context against which the equality is being tested
+         * \return True if the contexts are equal, false otherwise
+         */
+        bool operator==(const Context& rhs) const {
+            if (else_expr != rhs.else_expr) {
+                return false;
+            }
+            if (endif_expr != rhs.endif_expr) {
+                return false;
+            }
+            if (if_expr.has_value() && !rhs.if_expr.has_value()) {
+                return false;
+            }
+            if (!if_expr.has_value() && rhs.if_expr.has_value()) {
+                return false;
+            }
+            if (!if_expr.has_value() && !rhs.if_expr.has_value()) {
+                return true;
+            }
+
+            auto if_expr_val = if_expr.value();
+            auto rhs_if_expr_val = rhs.if_expr.value();
+            return if_expr_val.first == rhs_if_expr_val.first;
+        }
+
+        /**
+         * \brief Increments conditional locations by \a i
+         *
+         * \param i Non-negative integer
+         */
+        void inc_locs(idx i) {
+            if (if_expr.has_value()) {
+                if_expr.value().first += i;
+            }
+            if (else_expr.has_value()) {
+                else_expr.value() += i;
+            }
+            if (endif_expr.has_value()) {
+                endif_expr.value() += i;
+            }
+        }
+
+        /**
+         * \brief Increments dit shift by \a i
+         *
+         * \param i Non-negative integer
+         */
+        void inc_dit_shift(idx i) { dit_shift += i; }
+
+        /**
+         * \brief qpp::IDisplay::display() override
+         *
+         * \param os Output stream passed by reference
+         * \return Reference to the output stream
+         */
+        std::ostream& display(std::ostream& os) const override {
+            if (if_expr.has_value()) {
+                os << "IF: " << if_expr.value().first << ' ';
+                os << "addr. " << std::addressof(if_expr.value().second);
+            }
+            if (else_expr.has_value()) {
+                os << ", ELSE: " << else_expr.value();
+            }
+            if (endif_expr.has_value()) {
+                os << ", ENDIF: " << endif_expr.value();
+            }
+
+            return os;
+        }
     };
+
     Type condition_type_ = Type::NONE; ///< condition type
     Context ctx_;
-
-    /**
-     * \brief Extraction operator overload for
-     * qpp::internal::QCircuitConditionalStep::Context class
-     *
-     * \param os Output stream passed by reference
-     * \param measure_type qpp::internal::QCircuitConditionalStep::Context
-     *
-     * \return Reference to the output stream
-     */
-    friend std::ostream& operator<<(std::ostream& os, const Context& ctx) {
-        if (ctx.if_expr.has_value()) {
-            os << "IF: " << ctx.if_expr.value().first << ' ';
-            os << "addr. " << std::addressof(ctx.if_expr.value().second);
-        }
-        if (ctx.else_expr.has_value()) {
-            os << ", ELSE: " << ctx.else_expr.value();
-        }
-        if (ctx.endif_expr.has_value()) {
-            os << ", ENDIF: " << ctx.endif_expr.value();
-        }
-
-        return os;
-    }
 
     /**
      * \brief Default constructor
@@ -139,6 +192,7 @@ struct QCircuitConditionalStep : IDisplay {
      *
      * \tparam Func Conditional functor
      * \param condition_type Measurement type
+     * \param ctx Context
      */
     explicit QCircuitConditionalStep(Type condition_type, Context ctx)
         : condition_type_{condition_type}, ctx_(std::move(ctx)) {}
@@ -151,9 +205,9 @@ struct QCircuitConditionalStep : IDisplay {
      * \return True if the qpp::internal::QCircuitConditionalStep(s) are equal,
      * false otherwise
      */
-    // FIXME:
     bool operator==(const QCircuitConditionalStep& rhs) const noexcept {
-        return std::tie(rhs.condition_type_) == std::tie(condition_type_);
+        return std::tie(rhs.condition_type_, rhs.ctx_) ==
+               std::tie(condition_type_, ctx_);
     }
 
     /**